requirements.txt
torch
torchvision
tensorflow
opencv-python
scikit-learn
pyyaml
tqdm
colorama
pytest
